make it easy to porting to more picorv32 platform
